Move removeSoftDeletedBALs function & use Date-fns
