+++ lib/sync/index.js
@@ -254,17 +254,4 @@ async function syncOutdated() {
   }
 }
 
-async function removeSoftDeletedBALs() {
-  const now = new Date()
-  const deleteTime = new Date()
-
-  deleteTime.setMonth(now.getMonth() - 12)
-
-  const softDeletedBALs = await mongo.db.collection('bases_locales').deleteMany({
-    _deleted: {$lt: deleteTime}
-  })
-
-  console.log(`\n${softDeletedBALs.deletedCount} BALs supprimées définitivement\n`)
-}
-
-module.exports = {exec, pause, resume, detectOutdated, detectConflict, syncOutdated, removeSoftDeletedBALs}
+module.exports = {exec, pause, resume, detectOutdated, detectConflict, syncOutdated}
